requirements.txt
jax
flax
optax
ml-collections
tensorflow
tensorflow-datasets
Pillow
absl-py
tensorboard